refactor: move vfn controller into module
